requirements.txt
altair
pandas
streamlit
requests
supabase
python-dotenv